✨: Change website to load layout from API
The website is now able to communicate with a back-end API, which loads the layouts from a DynamoDB database. Currently it is assuming there is only one layout received, which is the current layout. A timestamp is stored with the layout however, which could be used to sort layouts in the future.